% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target.malt.rna.mat}
\alias{target.malt.rna.mat}
\title{Single cell RNA-sequencing (scRNA-seq) target subset of the 10X Genomics MALT counts.}
\format{
A scRNA-seq counts matrix of \code{dgCMatrix-class} from the \code{Matrix} package with 4000 cells and 33538 genes.
}
\source{
<https://www.10xgenomics.com/resources/datasets/10-k-cells-from-a-malt-tumor-gene-expression-and-cell-surface-protein-3-standard-3-0-0>
}
\usage{
target.malt.rna.mat
}
\description{
A random subset of joint scRNA-seq/CITE-seq 10X Genomics human extranodal marginal zone B-cell tumor/mucosa-associated lymphoid tissue (MALT) target data. See the \code{dataProcessing.R} file from \code{data-raw} folder for code to recreate data subset.
}
\keyword{datasets}
